Fix agent text comment review findings
- CSS: move ::highlight rules after @import/@plugin (CSS spec requires
  imports first; later imports could be silently dropped)
- Timestamp: make created_at BigInteger + microseconds to match updated_at
  (was Integer seconds vs BigInteger microseconds — 6 orders of magnitude
  mismatch)
- Re-anchoring: search with increment 1 instead of selected_text.length so
  overlapping repeated text matches are not skipped
- Shared send state: lift isSending/sentBatchIds to AppShell so desktop and
  mobile panels can't duplicate-send during viewport transitions
- Cross-session draft: capture conversationId in saveDraft and only activate
  the comment if still on the same session
- Session switch: close mobile comments drawer and reset send state on
  conversationId change (was only clearing pending/active)
- Query error: show explicit error + retry instead of empty state when
  loading comments fails
- Active comment: clear activeCommentId when deleting the active comment
- Memoize agentTextCommentsUI to prevent unnecessary context re-renders

+++ omnigent/db/db_models.py
@@ -1201,7 +1201,7 @@ class SqlAgentTextComment(ConversationBase):
     prefix_context: Mapped[str] = mapped_column(CompressedText)
     suffix_context: Mapped[str] = mapped_column(CompressedText)
     body: Mapped[str] = mapped_column(CompressedText)
-    created_at: Mapped[int] = mapped_column(Integer)
+    created_at: Mapped[int] = mapped_column(BigInteger)
     updated_at: Mapped[int] = mapped_column(BigInteger)
 
     __table_args__ = (
